+++ workspaces/src/types/mod.rs
@@ -206,10 +206,10 @@ pub struct AccessKey {
     /// NOTE: In some cases the access key needs to be recreated. If the new access key reuses the
     /// same public key, the nonce of the new access key should be equal to the nonce of the old
     /// access key. It's required to avoid replaying old transactions again.
-    nonce: Nonce,
+    pub nonce: Nonce,
 
     /// Defines permissions for this access key.
-    permission: AccessKeyPermission,
+    pub permission: AccessKeyPermission,
 }
 
 impl AccessKey {
@@ -236,9 +236,24 @@ impl AccessKey {
     }
 }
 
+/// Similar to an [`AccessKey`], but also has the [`PublicKey`] associated with it.
+pub struct AccessKeyInfo {
+    pub public_key: PublicKey,
+    pub access_key: AccessKey,
+}
+
+impl From<near_primitives::views::AccessKeyInfoView> for AccessKeyInfo {
+    fn from(view: near_primitives::views::AccessKeyInfoView) -> Self {
+        Self {
+            public_key: PublicKey(view.public_key),
+            access_key: view.access_key.into(),
+        }
+    }
+}
+
 /// Defines permissions for AccessKey
 #[derive(Clone, Debug)]
-enum AccessKeyPermission {
+pub enum AccessKeyPermission {
     FunctionCall(FunctionCallPermission),
 
     /// Grants full access to the account.
@@ -251,25 +266,25 @@ enum AccessKeyPermission {
 /// It also restrict the account ID of the receiver for this function call.
 /// It also can restrict the method name for the allowed function calls.
 #[derive(Clone, Debug)]
-struct FunctionCallPermission {
+pub struct FunctionCallPermission {
     /// Allowance is a balance limit to use by this access key to pay for function call gas and
     /// transaction fees. When this access key is used, both account balance and the allowance is
     /// decreased by the same value.
     /// `None` means unlimited allowance.
     /// NOTE: To change or increase the allowance, the old access key needs to be deleted and a new
     /// access key should be created.
-    allowance: Option<Balance>,
+    pub allowance: Option<Balance>,
 
     // This isn't an AccountId because already existing records in testnet genesis have invalid
     // values for this field (see: https://github.com/near/nearcore/pull/4621#issuecomment-892099860)
     // we accomodate those by using a string, allowing us to read and parse genesis.
     /// The access key only allows transactions with the given receiver's account id.
-    receiver_id: String,
+    pub receiver_id: String,
 
     /// A list of method names that can be used. The access key only allows transactions with the
     /// function call of one of the given method names.
     /// Empty list means any method name can be used.
-    method_names: Vec<String>,
+    pub method_names: Vec<String>,
 }
 
 impl From<AccessKey> for near_primitives::account::AccessKey {
@@ -291,5 +306,53 @@ impl From<AccessKey> for near_primitives::account::AccessKey {
                 }
             },
         }
+    }
+}
+
+impl From<near_primitives::views::AccessKeyView> for AccessKey {
+    fn from(access_key: near_primitives::views::AccessKeyView) -> Self {
+        Self {
+            nonce: access_key.nonce,
+            permission: match access_key.permission {
+                near_primitives::views::AccessKeyPermissionView::FunctionCall {
+                    allowance,
+                    receiver_id,
+                    method_names,
+                } => AccessKeyPermission::FunctionCall(FunctionCallPermission {
+                    allowance,
+                    receiver_id,
+                    method_names,
+                }),
+                near_primitives::views::AccessKeyPermissionView::FullAccess => {
+                    AccessKeyPermission::FullAccess
+                }
+            },
+        }
+    }
+}
+
+/// Finality of a transaction or block in which transaction is included in. For more info
+/// go to the [NEAR finality](https://docs.near.org/docs/concepts/transaction#finality) docs.
+#[derive(Debug)]
+#[non_exhaustive]
+pub enum Finality {
+    /// Optimistic finality. The latest block recorded on the node that responded to our query
+    /// (<1 second delay after the transaction is submitted).
+    Optimistic,
+    /// Near-final finality. Similiarly to `Final` finality, but delay should be roughly 1 second.
+    DoomSlug,
+    /// Final finality. The block that has been validated on at least 66% of the nodes in the
+    /// network. (At max, should be 2 second delay after the transaction is submitted.)
+    Final,
+}
+
+impl From<Finality> for near_primitives::types::BlockReference {
+    fn from(value: Finality) -> Self {
+        let value = match value {
+            Finality::Optimistic => near_primitives::types::Finality::None,
+            Finality::DoomSlug => near_primitives::types::Finality::DoomSlug,
+            Finality::Final => near_primitives::types::Finality::Final,
+        };
+        value.into()
     }
 }