+++ workspaces/tests/patch_state.rs
@@ -34,7 +34,7 @@ async fn view_status_state(
         .await?
         .into_result()?;
 
-    let mut state_items = contract.view_state(None).await?;
+    let mut state_items = contract.view_state().await?;
     let state = state_items
         .remove(b"STATE".as_slice())
         .ok_or_else(|| anyhow::anyhow!("Could not retrieve STATE"))?;
@@ -75,15 +75,10 @@ async fn test_patch_state() -> anyhow::Result<()> {
         .await?;
 
     let status: String = worker
-        .view(
-            &contract_id,
-            "get_status",
-            json!({
-                "account_id": "alice.near",
-            })
-            .to_string()
-            .into_bytes(),
-        )
+        .view(&contract_id, "get_status")
+        .args_json(json!({
+            "account_id": "alice.near",
+        }))
         .await?
         .json()?;